Added calorie tracker to model and admin page, and also created py tests for model
