+++ Bigfitproject/Bigfit/admin.py
@@ -7,10 +7,15 @@ class WeightTrackerInline(admin.TabularInline):
     ordering = ('record_date',)
 
 
+class CalorieTrackerInline(admin.TabularInline):
+    model = CalorieTracker
+    ordering = ('record_date',)
+
+
 class UserAdmin(admin.ModelAdmin):
     list_display = ['last_name', 'first_name', 'username', 'password','target_weight', 'feet'
                     , 'inches', 'date_of_birth', 'gender', 'zip_code']
-    inlines = [WeightTrackerInline]
+    inlines = [WeightTrackerInline, CalorieTrackerInline]
 
 
 admin.site.register(User, UserAdmin)
@@ -21,3 +26,10 @@ class WeightTrackerAdmin(admin.ModelAdmin):
 
 
 admin.site.register(WeightTracker, WeightTrackerAdmin)
+
+
+class CalorieTrackerAdmin(admin.ModelAdmin):
+    list_display = ['user', 'record_date', 'calories']
+
+
+admin.site.register(CalorieTracker, CalorieTrackerAdmin)
